Colorise path of cloud component
